+++ priv/repo/migrations/20190718055634_add_manage_quality_rule_implementation_permission.exs
@@ -1,0 +1,18 @@
+defmodule TdAuth.Repo.Migrations.AddManageQualityRuleImplementationPermission do
+  use Ecto.Migration
+
+  alias TdAuth.Permissions
+
+  @permissions [
+    "manage_quality_rule_implementations"
+  ]
+
+  def change do
+    @permissions
+      |> Enum.each(&load_permission/1)
+  end
+
+  defp load_permission(name) do
+    Permissions.create_permission(%{name: name})
+  end
+end