+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-1775] Permission manage quality rule implementations
+- [TD-1776] Permission view_quality_rule
+
 ## [3.2.0] 2019-07-24
 
 ### Changed